requirements.txt
asgiref==3.8.1
colorama==0.4.6
decorator==5.2.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.3.1
opencv-python==4.11.0.86
pillow==11.2.1
proglog==0.1.12
PyJWT==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.2


